% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethodHapsinator.R
\name{plot.hapsinatoR}
\alias{plot.hapsinatoR}
\title{Plot a hapsinatoR object}
\usage{
\method{plot}{hapsinatoR}(x)
}
\arguments{
\item{x}{An object of class \code{hapsinatoR}, typically returned by \code{hapsinatoR()}.}
}
\value{
A ggplot2 object.
}
\description{
Creates a tile plot of the candy shelf using ggplot2.
}
