% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethodHapsinator.R
\name{plot_highlighted_grid}
\alias{plot_highlighted_grid}
\title{plot hapsinator}
\usage{
plot_highlighted_grid(candyShelf)
}
\arguments{
\item{candyShelf}{}
}
\value{
print(pricePer100g)
}
\description{
plot hapsinator
}
