% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethodHapsinator.R
\name{print.hapsinatoR}
\alias{print.hapsinatoR}
\title{Print method for hapsinatoR objects}
\usage{
\method{print}{hapsinatoR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hapsinatoR}.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Prints a summary of the hapsinatoR object.
}
