% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMethodHapsinator.R
\name{summary.hapsinatoR}
\alias{summary.hapsinatoR}
\title{Summary method for hapsinatoR objects}
\usage{
\method{summary}{hapsinatoR}(object, ...)
}
\arguments{
\item{object}{An object of class \code{hapsinatoR}, typically created by the \code{hapsinatoR()} function.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Returns a summary on the hapsinatoR object
}
\description{
Prints a concise summary of a \code{hapsinatoR} object, including the number of selected candies (haps),
total weight to mix, number of unique candies, and coordinates of the candies to grab.
}
