% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.hapsinatoR.R
\name{validate_hapsinatoR}
\alias{validate_hapsinatoR}
\title{Validator for hapsinatoR object}
\usage{
validate_hapsinatoR(x)
}
\arguments{
\item{x}{Object to be checked if it is of class \code{"hapsinatoR"}}
}
\value{
Returns the object if it is of class \code{"hapsinatoR"} otherwise, throws an error.
}
\description{
Validator for hapsinatoR object
}
\keyword{internal}
