% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapsinator.R
\name{hapsinator}
\alias{hapsinator}
\title{hapsinator}
\usage{
hapsinator(pricePer100g, snollerMoney, numRows, numCols, handfulSize)
}
\arguments{
\item{pricePer100g}{price of 100g of candy}

\item{snollerMoney}{money to spend on candy}

\item{numRows}{number of rows in the candy shelf}

\item{numCols}{number of columns in the candy shelf}

\item{handfulSize}{handful size (small, medium or large)}
}
\value{
print(pricePer100g)
}
\description{
hapsinator
}
\examples{
bla
}
