% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapsinator.R
\name{hapsinatoR}
\alias{hapsinatoR}
\title{CandyHapsinator}
\usage{
hapsinatoR(
  pricePer100g = NULL,
  snollerMoney = NULL,
  numRows = NULL,
  numCols = NULL,
  handfulSize = c("medium", "small", "large")
)
}
\arguments{
\item{pricePer100g}{price of 100g of candy}

\item{snollerMoney}{money to spend on candy}

\item{numRows}{number of rows in the candy shelf}

\item{numCols}{number of columns in the candy shelf}

\item{handfulSize}{handful size (small, medium or large)}
}
\value{
An object of class \sQuote{hapsinatoR}.
An object of \code{\link{class}} \code{"hapsinatoR"} is a list containing the following components
}
\description{
CandyHapsinator
}
